replaced all pixel types uses of unsigned, short, etc with explicit vxl_uint_32, vxl_int_16 as per vxl meeting agreement
Removed references to rgb pixels from vil2_pnm_image. I have not yet been persuaded
that the expense (in terms of lines of code to write/maintain) of supporting rgb pixels
on top of multi-plane,scalar-components in vil2_image_data is a worthwhile tradeoff.
It wil be trivial to provide support for this at the vil2_image_view level, and trivial
to use the support there play with rgb images as much as you like.


git-svn-id: https://vxl.svn.sourceforge.net/svnroot/vxl/trunk@6546 4b17d0e5-ac7a-4109-a8ed-6b3c4e86a14a