Converted classes using "operator bool" to instead use the "safe-bool" idiom.  The idea is that conversion to bool allows unexpected comparisons between unrelated types to compile without error because both sides can convert to bool.
The "safe-bool" idiom looks like this inside a class definition:

private:
  struct safe_bool_dummy { void dummy() {} };
  typedef void (safe_bool_dummy::* safe_bool)();
public:
  operator safe_bool () const
    { return (CONDITION)? &safe_bool_dummy::dummy : 0; }

Using "operator safe_bool" instead solves this problem because the
"safe_bool" type is defined as a pointer to a member of a private type
inside each class using the idiom.  This pointer can be used in a
boolean context but will not allow accidental comparison of unrelated
types to compile.


git-svn-id: https://vxl.svn.sourceforge.net/svnroot/vxl/trunk@13909 4b17d0e5-ac7a-4109-a8ed-6b3c4e86a14a