Changes made to work around bcc compiler errors:
Removed 'using' statements for partially overloaded value() methods in typed_array_field and typed_scalar_field subclasses. I don't believe they are needed, as its OK for the other overloads to be hidden in the subclass.

Also renamed one of the read and write methods in field_formatter to avoid method overloading.


git-svn-id: https://vxl.svn.sourceforge.net/svnroot/vxl/trunk@16560 4b17d0e5-ac7a-4109-a8ed-6b3c4e86a14a