Encountered a strange error where JPEG library aborted the program because it cannot write temporary data into temporary directory on Windows.
It turned out there was relating discussion in this post (from different software project):
http://trac.osgeo.org/gdal/ticket/1795

To fix this issue and to boost the performance of JPEG compression and decompression, I introduced a pre-defined macro VXL_ADDRESS_BITS that specifies the build environment. This macro value is determined from the size of size_t.  If it is 32-bit build environment, I increased the memory to be used by JPEG lib to 300Mb.  If it is 64 bit, it is increased to 1GB. 

git-svn-id: https://vxl.svn.sourceforge.net/svnroot/vxl/trunk@29468 4b17d0e5-ac7a-4109-a8ed-6b3c4e86a14a